Made forms a bit more error-friendly.
I went through and added 'required' to form inputs, in order to make it
so that the form doesn't send if there is an empty field. I used a
little JS to make it so that if passwords don’t match, in admin
registration, and user registration, it won't send the form through and
will leave the users inputs in the input fields. I added minimum and
maximum lengths in the post idea area.