Made forms more error-friendly
I added 'required' to (almost) all of the form fields, i added
"minlength' to the admin password creation field, register user password
field, and change password field, as well as the title field and
description field in the post new idea area. I used a little bit of JS
to check if passwords and re-type passwords match. I also fixed a small
typo in 'register.php' that made it so that if an email already existed,
it wouldn't display the error message, I changed 'exits' to 'exists'.